app/models/gold_price.py
import requests
from bs4 import BeautifulSoup
from datetime import datetime

class GoldPrice:
    @staticmethod
    def fetch_latest_prices():
        try:
            response = requests.get(Config.PUBLIC_GOLD_URL)
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # Extract gold prices from Public Gold website
            prices = {
                'gram_1': self._extract_price(soup, '1 gram'),
                'gram_5': self._extract_price(soup, '5 gram'),
                'gram_10': self._extract_price(soup, '10 gram'),
                'updated_at': datetime.now()
            }
            return prices
        except Exception as e:
            print(f"Error fetching gold prices: {e}")
            return None

    @staticmethod
    def _extract_price(soup, weight):
        # Implement specific price extraction logic based on Public Gold's HTML structure
        pass