requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
requests==2.26.0
Pillow==9.1.0
APScheduler==3.8.1
beautifulsoup4==4.9.3
python-dotenv==0.19.0
Werkzeug==2.2.2 